+++ CHANGELOG.md
@@ -9,3 +9,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 Release date `UNRELEASED`
 
 - (feat api): add `csp/policy-header-fn` function
+- (feat api): add `csp/random-nonce` function